+++ deps/vecmath/src/Matrix3f.cpp
@@ -9,7 +9,7 @@
 #include "Quat4f.h"
 #include "Vector3f.h"
 
-Matrix3f::Matrix3f( float fill )
+Matrix3f::Matrix3f( double fill )
 {
 	for( int i = 0; i < 9; ++i )
 	{
@@ -17,9 +17,9 @@ Matrix3f::Matrix3f( float fill )
 	}
 }
 
-Matrix3f::Matrix3f( float m00, float m01, float m02,
-				   float m10, float m11, float m12,
-				   float m20, float m21, float m22 )
+Matrix3f::Matrix3f( double m00, double m01, double m02,
+				   double m10, double m11, double m12,
+				   double m20, double m21, double m22 )
 {
 	m_elements[ 0 ] = m00;
 	m_elements[ 1 ] = m10;
@@ -52,24 +52,24 @@ Matrix3f::Matrix3f( const Vector3f& v0, const Vector3f& v1, const Vector3f& v2, 
 
 Matrix3f::Matrix3f( const Matrix3f& rm )
 {
-	memcpy( m_elements, rm.m_elements, 9 * sizeof( float ) );
+	memcpy( m_elements, rm.m_elements, 9 * sizeof( double ) );
 }
 
 Matrix3f& Matrix3f::operator = ( const Matrix3f& rm )
 {
 	if( this != &rm )
 	{
-		memcpy( m_elements, rm.m_elements, 9 * sizeof( float ) );
+		memcpy( m_elements, rm.m_elements, 9 * sizeof( double ) );
 	}
 	return *this;
 }
 
-const float& Matrix3f::operator () ( int i, int j ) const
+const double& Matrix3f::operator () ( int i, int j ) const
 {
 	return m_elements[ j * 3 + i ];
 }
 
-float& Matrix3f::operator () ( int i, int j )
+double& Matrix3f::operator () ( int i, int j )
 {
 	return m_elements[ j * 3 + i ];
 }
@@ -138,7 +138,7 @@ void Matrix3f::setSubmatrix2x2( int i0, int j0, const Matrix2f& m )
 	}
 }
 
-float Matrix3f::determinant() const
+double Matrix3f::determinant() const
 {
 	return Matrix3f::determinant3x3
 	(
@@ -148,33 +148,33 @@ float Matrix3f::determinant() const
 	);
 }
 
-Matrix3f Matrix3f::inverse( bool* pbIsSingular, float epsilon ) const
+Matrix3f Matrix3f::inverse( bool* pbIsSingular, double epsilon ) const
 {
-	float m00 = m_elements[ 0 ];
-	float m10 = m_elements[ 1 ];
-	float m20 = m_elements[ 2 ];
+	double m00 = m_elements[ 0 ];
+	double m10 = m_elements[ 1 ];
+	double m20 = m_elements[ 2 ];
 
-	float m01 = m_elements[ 3 ];
-	float m11 = m_elements[ 4 ];
-	float m21 = m_elements[ 5 ];
+	double m01 = m_elements[ 3 ];
+	double m11 = m_elements[ 4 ];
+	double m21 = m_elements[ 5 ];
 
-	float m02 = m_elements[ 6 ];
-	float m12 = m_elements[ 7 ];
-	float m22 = m_elements[ 8 ];
+	double m02 = m_elements[ 6 ];
+	double m12 = m_elements[ 7 ];
+	double m22 = m_elements[ 8 ];
 
-	float cofactor00 =  Matrix2f::determinant2x2( m11, m12, m21, m22 );
-	float cofactor01 = -Matrix2f::determinant2x2( m10, m12, m20, m22 );
-	float cofactor02 =  Matrix2f::determinant2x2( m10, m11, m20, m21 );
+	double cofactor00 =  Matrix2f::determinant2x2( m11, m12, m21, m22 );
+	double cofactor01 = -Matrix2f::determinant2x2( m10, m12, m20, m22 );
+	double cofactor02 =  Matrix2f::determinant2x2( m10, m11, m20, m21 );
 
-	float cofactor10 = -Matrix2f::determinant2x2( m01, m02, m21, m22 );
-	float cofactor11 =  Matrix2f::determinant2x2( m00, m02, m20, m22 );
-	float cofactor12 = -Matrix2f::determinant2x2( m00, m01, m20, m21 );
+	double cofactor10 = -Matrix2f::determinant2x2( m01, m02, m21, m22 );
+	double cofactor11 =  Matrix2f::determinant2x2( m00, m02, m20, m22 );
+	double cofactor12 = -Matrix2f::determinant2x2( m00, m01, m20, m21 );
 
-	float cofactor20 =  Matrix2f::determinant2x2( m01, m02, m11, m12 );
-	float cofactor21 = -Matrix2f::determinant2x2( m00, m02, m10, m12 );
-	float cofactor22 =  Matrix2f::determinant2x2( m00, m01, m10, m11 );
+	double cofactor20 =  Matrix2f::determinant2x2( m01, m02, m11, m12 );
+	double cofactor21 = -Matrix2f::determinant2x2( m00, m02, m10, m12 );
+	double cofactor22 =  Matrix2f::determinant2x2( m00, m01, m10, m11 );
 
-	float determinant = m00 * cofactor00 + m01 * cofactor01 + m02 * cofactor02;
+	double determinant = m00 * cofactor00 + m01 * cofactor01 + m02 * cofactor02;
 	
 	bool isSingular = ( fabs( determinant ) < epsilon );
 	if( isSingular )
@@ -192,7 +192,7 @@ Matrix3f Matrix3f::inverse( bool* pbIsSingular, float epsilon ) const
 			*pbIsSingular = false;
 		}
 
-		float reciprocalDeterminant = 1.0f / determinant;
+		double reciprocalDeterminant = 1.0f / determinant;
 
 		return Matrix3f
 		(
@@ -205,7 +205,7 @@ Matrix3f Matrix3f::inverse( bool* pbIsSingular, float epsilon ) const
 
 void Matrix3f::transpose()
 {
-	float temp;
+	double temp;
 
 	for( int i = 0; i < 2; ++i )
 	{
@@ -232,7 +232,7 @@ Matrix3f Matrix3f::transposed() const
 	return out;
 }
 
-Matrix3f::operator float* ()
+Matrix3f::operator double* ()
 {
 	return m_elements;
 }
@@ -246,9 +246,9 @@ void Matrix3f::print()
 }
 
 // static
-float Matrix3f::determinant3x3( float m00, float m01, float m02,
-							   float m10, float m11, float m12,
-							   float m20, float m21, float m22 )
+double Matrix3f::determinant3x3( double m00, double m01, double m02,
+							   double m10, double m11, double m12,
+							   double m20, double m21, double m22 )
 {
 	return
 		(
@@ -284,10 +284,10 @@ Matrix3f Matrix3f::identity()
 
 
 // static
-Matrix3f Matrix3f::rotateX( float radians )
+Matrix3f Matrix3f::rotateX( double radians )
 {
-	float c = cos( radians );
-	float s = sin( radians );
+	double c = cos( radians );
+	double s = sin( radians );
 
 	return Matrix3f
 	(
@@ -298,10 +298,10 @@ Matrix3f Matrix3f::rotateX( float radians )
 }
 
 // static
-Matrix3f Matrix3f::rotateY( float radians )
+Matrix3f Matrix3f::rotateY( double radians )
 {
-	float c = cos( radians );
-	float s = sin( radians );
+	double c = cos( radians );
+	double s = sin( radians );
 
 	return Matrix3f
 	(
@@ -312,10 +312,10 @@ Matrix3f Matrix3f::rotateY( float radians )
 }
 
 // static
-Matrix3f Matrix3f::rotateZ( float radians )
+Matrix3f Matrix3f::rotateZ( double radians )
 {
-	float c = cos( radians );
-	float s = sin( radians );
+	double c = cos( radians );
+	double s = sin( radians );
 
 	return Matrix3f
 	(
@@ -326,7 +326,7 @@ Matrix3f Matrix3f::rotateZ( float radians )
 }
 
 // static
-Matrix3f Matrix3f::scaling( float sx, float sy, float sz )
+Matrix3f Matrix3f::scaling( double sx, double sy, double sz )
 {
 	return Matrix3f
 	(
@@ -337,7 +337,7 @@ Matrix3f Matrix3f::scaling( float sx, float sy, float sz )
 }
 
 // static
-Matrix3f Matrix3f::uniformScaling( float s )
+Matrix3f Matrix3f::uniformScaling( double s )
 {
 	return Matrix3f
 	(
@@ -348,16 +348,16 @@ Matrix3f Matrix3f::uniformScaling( float s )
 }
 
 // static
-Matrix3f Matrix3f::rotation( const Vector3f& rDirection, float radians )
+Matrix3f Matrix3f::rotation( const Vector3f& rDirection, double radians )
 {
 	Vector3f normalizedDirection = rDirection.normalized();
 	
-	float cosTheta = cos( radians );
-	float sinTheta = sin( radians );
+	double cosTheta = cos( radians );
+	double sinTheta = sin( radians );
 
-	float x = normalizedDirection.x();
-	float y = normalizedDirection.y();
-	float z = normalizedDirection.z();
+	double x = normalizedDirection.x();
+	double y = normalizedDirection.y();
+	double z = normalizedDirection.z();
 
 	return Matrix3f
 		(
@@ -372,18 +372,18 @@ Matrix3f Matrix3f::rotation( const Quat4f& rq )
 {
 	Quat4f q = rq.normalized();
 
-	float xx = q.x() * q.x();
-	float yy = q.y() * q.y();
-	float zz = q.z() * q.z();
+	double xx = q.x() * q.x();
+	double yy = q.y() * q.y();
+	double zz = q.z() * q.z();
 
-	float xy = q.x() * q.y();
-	float zw = q.z() * q.w();
+	double xy = q.x() * q.y();
+	double zw = q.z() * q.w();
 
-	float xz = q.x() * q.z();
-	float yw = q.y() * q.w();
+	double xz = q.x() * q.z();
+	double yw = q.y() * q.w();
 
-	float yz = q.y() * q.z();
-	float xw = q.x() * q.w();
+	double yz = q.y() * q.z();
+	double xw = q.x() * q.w();
 
 	return Matrix3f
 		(
